# app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Instagram Clone API",
        "description": "API do projeto pessoal - clone do Instagram feito em Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/stories": {
            "get": {
                "tags": [
                    "Stories"
                ],
                "summary": "Feed de stories (usuários seguidos)",
                "operationId": "3b7ca25d3f2902333dc031e052d6f3f9",
                "responses": {
                    "200": {
                        "description": "Stories ativas agrupadas por usuário"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Stories"
                ],
                "summary": "Criar uma story",
                "operationId": "9bd3f80317a4d0b0d7a4f0de3e8108a3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "media"
                                ],
                                "properties": {
                                    "media": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Story criada"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/stories/mine": {
            "get": {
                "tags": [
                    "Stories"
                ],
                "summary": "Listar as próprias stories ativas",
                "operationId": "a93c9e23fb0fc2baba99680bd98620da",
                "responses": {
                    "200": {
                        "description": "Lista de stories do usuário"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/stories/{story}": {
            "delete": {
                "tags": [
                    "Stories"
                ],
                "summary": "Excluir uma story",
                "operationId": "217ef5d96a7be886f874397ad79c516b",
                "parameters": [
                    {
                        "name": "story",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Story excluída"
                    },
                    "403": {
                        "description": "Não é o dono da story"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/users": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Listar/buscar usuários",
                "operationId": "7e2f5799553ec93c9c43e2a58490d447",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Busca por nome ou username",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de usuários"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/users/suggestions": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Sugestões de usuários para seguir",
                "operationId": "53d0848e159311a49b0e452f5d2ee493",
                "responses": {
                    "200": {
                        "description": "Até 5 usuários aleatórios que você ainda não segue"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/users/recommended": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Recomendados para seguir",
                "operationId": "dd46cea11c0504e4f2deabfbf100bd25",
                "responses": {
                    "200": {
                        "description": "Usuários que curtiram os mesmos posts que você (ou que curtiram seus posts), ordenados por afinidade, excluindo quem você já segue"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/users/{user}": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Ver perfil de um usuário",
                "operationId": "0b9800c6a2369ce94c1efa162edfa374",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do perfil, contagens e posts"
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Editar o próprio perfil",
                "operationId": "caaec06bf33a577d630c028222869ee6",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Maria Silva"
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "mariasilva"
                                    },
                                    "bio": {
                                        "type": "string",
                                        "example": "Nova bio"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Perfil atualizado"
                    },
                    "403": {
                        "description": "Não é o dono do perfil"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Excluir a própria conta",
                "operationId": "a4e18f8517afcc062825920ffb6cd762",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Conta excluída (soft delete)"
                    },
                    "403": {
                        "description": "Não é o dono da conta"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/users/{user}/avatar": {
            "post": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Trocar foto de perfil",
                "operationId": "ecf262198917fd26404de6685b4235c3",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "avatar"
                                ],
                                "properties": {
                                    "avatar": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Avatar atualizado"
                    },
                    "403": {
                        "description": "Não é o dono do perfil"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}